Initial attempt at using emotion styles
