+++ src/renderer/components/SoundFileInfo.tsx
@@ -2,7 +2,8 @@ import React from "react";
 
 import { transparentize } from "polished";
 import { Icon } from "semantic-ui-react";
-import styled from "styled-components";
+
+import styled from "@emotion/styled";
 import { Labelled } from "./Labelled";
 
 const Container = styled.div`
